Throw exception to terminate if input is invalid
The fail should cause immediate termination of an activity if an error
is detected. Hence fail method should throw an exception instead of returning
a new one to the caller. The callers delegate the responsibility of generating
an exception uniformly.

If the exception is returned, the processing continues if the caller
does not act on  the returned exception, which causes unhandled errors later.